MainProductSlider converted into object
